Update compareMS2.c
Added spectral angle (SA) as alternative measure for comparing two tandem mass spectra. While this is just a simple transformation of the dot product, it stretches out the best matches over a wider range, making it easier to distinguish good from great matches in visualizations. Note that the cutoff probably also need adjustment when replacing the dot product with SA.